Split setup screen up into sections (and use a SectionContainer)
